Add some documentation

+++ app/src/main/java/edu/ucsb/cs/cs190i/ihungry/ihungryapp/RequestBuilder.java
@@ -42,11 +42,11 @@ public class RequestBuilder {
     }
 
     /**
-     * Sets the Request's radius
+     * Sets the Request's radius in meters
      * @param radius
      * @return RequestBuilder
      */
-    public RequestBuilder radius(float radius) {
+    public RequestBuilder radius(double radius) {
         this.radius = radius;
         return this;
     }
